+++ src/components/Carousel/Carousel.js
@@ -1,6 +1,8 @@
 import React, {
+  forwardRef,
   Fragment,
   useCallback,
+  useImperativeHandle,
   useLayoutEffect,
   useRef,
   useState
@@ -32,7 +34,7 @@ import {
 import ReversedMap from '../../utils/ReversedMap';
 import { propTypes, defaultProps, getSettings } from './props';
 
-export const Carousel = (props) => {
+const GalleryCarousel = (props, ref) => {
   /* initialize references */
   const documentRef = useRef(document);
   const maximizedBackgroundRef = useRef(null);
@@ -207,6 +209,7 @@ export const Carousel = (props) => {
   const shouldCalibrateIndex = props.isLoop && nSlides > 1;
 
   const handleSwipeMoveX = (displacementX) => {
+    props.onSwipeMoveX(displacementX);
     // stop the timer for autoplay if there is a timer
     // should not use setIsPlaying(false) here since it will update the icon in the media button
     if (props.canAutoPlay) stopTimer();
@@ -311,11 +314,13 @@ export const Carousel = (props) => {
   isMaximizedRef.current = isMaximized;
 
   const handleSwipeMoveY = (displacementX, displacementY) => {
+    props.onSwipeMoveY(displacementX, displacementY);
     if (!props.shouldMinimizeOnSwipeDown) return;
     if (isMaximizedRef.current) applyTransitionY(displacementX, displacementY);
   };
 
   const handleSwipeEndDown = () => {
+    props.onSwipeEndDown();
     if (!props.shouldMinimizeOnSwipeDown) return;
     applyTransitionY(0, 0);
     setIsMaximized(() => false);
@@ -323,6 +328,7 @@ export const Carousel = (props) => {
   };
 
   const handleTap = () => {
+    props.onTap();
     if (isMaximizedRef.current && props.shouldMinimizeOnClick)
       setIsMaximized(() => false);
     else if (!isMaximizedRef.current && props.shouldMaximizeOnClick)
@@ -493,6 +499,19 @@ export const Carousel = (props) => {
     </>
   );
 
+  /* provide handlers for controlling the carousel in an imperative way */
+  useImperativeHandle(ref, () => ({
+    play: () => setIsPlaying(() => true),
+    pause: () => setIsPlaying(() => false),
+    toggleIsPlaying: () => setIsPlaying((isPlaying) => !isPlaying),
+    maximize: () => setIsMaximized(() => true),
+    minimize: () => setIsMaximized(() => false),
+    toggleIsMaximized: () => setIsMaximized((isMaximized) => !isMaximized),
+    goLeft: goLeft,
+    goRight: goRight,
+    goToIndex: goToIndex
+  }));
+
   return (
     <>
       {carouselPlaceholder}
@@ -539,5 +558,8 @@ export const Carousel = (props) => {
   );
 };
 
+export const Carousel = forwardRef(GalleryCarousel);
+
+Carousel.displayName = 'Carousel';
 Carousel.propTypes = propTypes;
 Carousel.defaultProps = defaultProps;